Controller and Model Update 1
First phase to add separation of concerns with a Controller and Model.
