+++ App_Code/ProjectStructs.cs
@@ -6,7 +6,10 @@ using System.Web;
 /// <summary>
 /// Summary description for ProjectStructs
 /// </summary>
-public class ProjectStructs
+/// 
+namespace ProjectStructs
+{
+public class DataStructs
 {
   public struct Client
   {
@@ -17,16 +20,33 @@ public class ProjectStructs
     public string email;
   }
 
+  public struct Appointment
+  {
+    public int id;
+    public string desiredDate;
+    public string startingTime;
+    public string endingTime;
+    public string services;
+    public int hairStylist;
+    public int hairLenght;
+    public int idCustomer;
+    public DateTime registrationDate;
+    public int registeredBy;
+    public bool cancelled;
+    public string cancellationReason;
+  }
+
   public struct TimeSlotsStartingAndEndingTimes
   {
     public TimeSpan serviceStartTime;
     public TimeSpan serviceEndTime;
   }
 
-	public ProjectStructs()
+	public DataStructs()
 	{
 		//
 		// TODO: Add constructor logic here
 		//
 	}
+}
 }
