+++ Web.config
@@ -34,7 +34,9 @@
         <add extension=".edmx" type="System.Data.Entity.Design.AspNet.EntityDesignerBuildProvider"/>
       </buildProviders>
     </compilation>
-    <customErrors mode="Off"/>
+    <customErrors mode="On" defaultRedirect="~/Errors/ErrorPage.aspx?handler=customErrors%20section%20-%20Web.config">
+      <error statusCode="404" redirect="~/Errors/ErrorPage.aspx?msg=404&amp;handler=customErrors%20section%20-%20Web.config"/>
+    </customErrors>  
     <httpRuntime targetFramework="4.5.1"/>
   </system.web>
   <entityFramework>
